FIX :: types 모듈, Layout 모듈 추가하여 수정
