+++ apps/internship/src/Router.tsx
@@ -1,46 +1,8 @@
-import {
-  BrowserRouter,
-  Routes,
-  Route,
-  Outlet,
-  useLocation,
-} from "react-router-dom";
-import { Header } from "./components/common/Header";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { Home } from "./pages/Home";
 import { JobStatus } from "./pages/JobStatus";
 import { Submitted } from "./pages/Submitted";
-
-export type IHeader = {
-  userType: "admin" | "user";
-  isLogin: boolean;
-};
-
-const ROUTE_USER_TYPES = {
-  "/": "admin",
-  "/job-status": "admin",
-  "/submitted": "user",
-} as const;
-
-const DEFAULT_USER_TYPE = "user";
-
-const getUserType = (pathname: string) => {
-  return (
-    ROUTE_USER_TYPES[pathname as keyof typeof ROUTE_USER_TYPES] ||
-    DEFAULT_USER_TYPE
-  );
-};
-
-const Layout = () => {
-  const location = useLocation();
-  const userType = getUserType(location.pathname);
-
-  return (
-    <>
-      <Header userType={userType} isLogin={false} />
-      <Outlet />
-    </>
-  );
-};
+import { Layout } from "./components/Layout";
 
 export const Router = () => {
   return (
